requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
numpy
google-generativeai
